+++ src/lib/update/ship.c
@@ -345,7 +345,7 @@ shiprepair(struct shpstr *ship, struct natstr *np,
     if (left > delta)
 	left = delta;
 
-    leftp = ((float)left / 100.0);
+    leftp = left / 100.0;
     memset(mvec, 0, sizeof(mvec));
     mvec[I_LCM] = lcm_needed = ldround(mp->m_lcm * leftp, 1);
     mvec[I_HCM] = hcm_needed = ldround(mp->m_hcm * leftp, 1);
